STRETCH : sessions data can now be stored in the database
